feat: refactor error handling by moving TableStorageError type to a separate file
