+++ packages/azure-sdk/src/data-tables/index.ts
@@ -1,0 +1,2 @@
+export { TableStorageError } from "./errors.js";
+export { TableClientWrapper } from "./data-table.wrapper.js";
